+++ packages/vinxi-server-functions/server-handler.js
@@ -10,7 +10,7 @@ export async function handleServerAction(event) {
 		invariant(typeof serverReference === "string", "Invalid server action");
 		// This is the client-side case
 		const [filepath, name] = serverReference.split("#");
-		const action = (await import.meta.env.MANIFEST["server"].chunks[filepath].import())[name];
+		const action = (await import.meta.env.MANIFEST[import.meta.env.ROUTER_NAME].chunks[filepath].import())[name];
 		const text = await new Promise((resolve) => {
 			const requestBody = [];
 			event.node.req.on("data", (chunks) => {
@@ -27,7 +27,7 @@ export async function handleServerAction(event) {
 			// Wait for any mutations
 			const response = await result;
 			event.node.res.setHeader("Content-Type", "application/json");
-			event.node.res.setHeader("Router", "server");
+			event.node.res.setHeader("Router", "server-fns");
 
 			return JSON.stringify(response ?? null);
 		} catch (x) {