+++ examples/solid/ssr/solid-start/CHANGELOG.md
@@ -4,6 +4,14 @@
 
 ### Patch Changes
 
+- Updated dependencies [a4512f5]
+  - vinxi@0.0.46
+  - @vinxi/solid-start@0.2.4
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [8d1f0dd]
   - vinxi@0.0.45
   - @vinxi/solid-start@0.2.4