+++ packages/vinxi-server-components/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @vinxi/plugin-references
 
+## 0.0.38
+
+### Patch Changes
+
+- Updated dependencies [a4512f5]
+  - vinxi@0.0.46
+  - @vinxi/plugin-directives@0.0.38
+
 ## 0.0.37
 
 ### Patch Changes